requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.28.1
python-multipart==0.0.18
stripe==7.10.0
beautifulsoup4==4.12.3
slowapi==0.1.9
